Day 6, part 2 optimization
The solution for part 2 in [5200033] was a brute force that just checked
for an obstacle in every position in the grid.  We didn't care about
whether or not it was "legal" to put an obstacle there:
	- It doesn't matter for our calculations if an obstacle is
	  already there.
	- I doesn't matter if the obstacle is placed on the guard,
	  because that would be his start position, and so he wouldn't
	  be able to walk a loop.

I realized that, instead of trying out placing an obstacle on all the
positions in the grid, we just need to check the places where the guard
can walk, which is calculated already in part 1.

This still looks very suboptimal, but it's about 4-5 times quicker than
previous solution (1.4s vs 6.5s)