Day 11, part 2
Of course it wasn't just `s/25/75`... I tried, but after consuming +70GB
of memory after running 8 minutes, I killed the process :-)

Thx @Jelle-S who said that it probably doesn't make a difference if you
process all stones x times or a each single stone x times, which sparked
the idea of blinking x times per stone (and the consequential recursion
+ caching of the results of the counting func)

Would be nice if Go had memoization (cf. Ptyhon's functools.cache
decorator) and there's probably a Go lib for that, but it isn't going to
be as simple to use as Python's and passing the cache map around isn't
that big of a deal, it could be a module-level variable even.