Day 19, part 1
I wanted to have a func CanMake() that would return a bool if a design
could be made, but that didn't take into account the different ways a
towel can be made... So I wrote it to find all the combinations with
memoizing the function.  I figured part 2 would be something along those
lines anyway.