Day 15: Straighten off-by-one compensation
I added the start position to the list of consecutive boxes, thinking it
would be easier to move them all together, but it was a bit a ugly, so I
moved the boxes and the robot separately.
Instead of not adding the robot to `consecutiveBoxes`, the quickest way
at that time was starting the loop over `consecutiveBoxes` start at 1.

This is now bothering me :-)