Introducing the `input` package
Going by 2 AoC's in Go, some ways of reading the puzzle's input return
time and time again.  This is by no means a package that covers all
possible ways, but I feel that it has some potential.
It uses bufio.Scanner to go through the input and makes use of functors
to process lines.  This covers most of the use cases I've seen, or at
least it provides some structure to process input in a reasonable custom
way.
The `TwoParts` will probably still result in a few lines more code
(although, I think that most of the times the functor definitions for
the processors are fairly simple one liners)

In the end, his package should replace the current `lib/file/file.go`
package.